+++ nisystemlink/clients/feeds/__init__.py
@@ -1,0 +1,3 @@
+from ._feeds_client import FeedsClient
+
+# flake8: noqa