+++ nisystemlink/clients/feeds/models/__init__.py
@@ -1,0 +1,7 @@
+from ._create_feed import CreateFeedRequest
+from ._feed import Feed
+from ._package import Package
+from ._platform import Platform
+from ._query_feeds import QueryFeedsResponse
+
+# flake8: noqa